refactor!: move load_all_events to more visible location
